+++ sdk/ai/azure-ai-agents/src/samples/java/com/azure/ai/agents/ReadmeSamples.java
@@ -27,8 +27,6 @@ import com.openai.models.responses.Response;
 import com.openai.models.responses.ResponseCreateParams;
 import com.openai.services.blocking.ConversationService;
 
-import java.util.Collections;
-
 public final class ReadmeSamples {
     public void readmeSamples() {
         String endpoint = "my-resource-url";
@@ -52,8 +50,8 @@ public final class ReadmeSamples {
 
         // BEGIN: com.azure.ai.agents.configure_agent_endpoint
         AgentEndpointConfig endpointConfig = new AgentEndpointConfig()
-            .setVersionSelector(new VersionSelector().setVersionSelectionRules(Collections.singletonList(
-                new FixedRatioVersionSelectionRule(100).setAgentVersion(agent.getVersion()))))
+            .setVersionSelector(new VersionSelector().setVersionSelectionRule(
+                new FixedRatioVersionSelectionRule(100).setAgentVersion(agent.getVersion())))
             .setProtocolConfiguration(new ProtocolConfiguration().setResponses(new ResponsesProtocolConfiguration()));
 
         agentsClient.updateAgentDetails(agent.getName(),